refactor(render): move RenderOnWeb trait to render module
